+++ server/impl/src/test/java/com/walmartlabs/concord/server/security/apikey/ApiKeyDaoTest.java
@@ -1,6 +1,7 @@
 package com.walmartlabs.concord.server.security.apikey;
 
 import com.walmartlabs.concord.server.AbstractDaoTest;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.security.SecureRandom;
@@ -8,6 +9,7 @@ import java.security.SecureRandom;
 import static org.junit.Assert.assertNotNull;
 import static org.mockito.Mockito.mock;
 
+@Ignore("requires a local DB instance")
 public class ApiKeyDaoTest extends AbstractDaoTest {
 
     @Test